refactor: split OpenAI preferences out of DayCountersStore
Address review feedback: revert the detekt.yml TooManyFunctions
threshold bump in favor of a real fix. DayCountersStore was doing
double duty - daily learning counters/policy plus unrelated OpenAI
settings - and three of its four consumers (AddWordViewModel,
GenerateAiTranslationUseCase, VocabularyFunctions) only ever touched
the OpenAI half. Extract those into a dedicated OpenAiPreferencesStore,
backed by the same underlying DataStore<Preferences> instance via a
new shared StudyPreferencesDataStore holder (DataStore disallows
multiple active instances for the same file, so the instance must be
shared rather than duplicated). This drops DayCountersStore back under
the function-count threshold without loosening it project-wide.

Also removes a stray leftover comment in
VocabularyRepositoryGetNextItemTest.kt.
