+++ app/src/main/java/com/procrastilearn/app/data/local/prefs/StudyPreferencesDataStore.kt
@@ -1,0 +1,27 @@
+package com.procrastilearn.app.data.local.prefs
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.dataStoreFile
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+import javax.inject.Singleton
+
+/**
+ * Owns the single DataStore<Preferences> instance backing "study_prefs.preferences_pb".
+ * DataStore disallows multiple active instances for the same file, so any class reading
+ * or writing these preferences must share this instance rather than creating its own.
+ */
+@Singleton
+class StudyPreferencesDataStore
+    @Inject
+    constructor(
+        @ApplicationContext ctx: Context,
+    ) {
+        val ds: DataStore<Preferences> =
+            PreferenceDataStoreFactory.create(
+                produceFile = { ctx.dataStoreFile("study_prefs.preferences_pb") },
+            )
+    }
